plugin interface for manila to backends
This commit contains the interface code to connect between the
manila charm and a backend plugin that will provide configuration
for the manila.

Plugin charms are subordinate because they need to be on the same
unit/machine as the manila file share service.

This interface also implements a 1-many interface and so contains code
to manage that.  That code will be abstracted back to charms.openstack
at some point.

Change-Id: Ied0ad014ab7b1d4778113b0d3f2bbae08075372e