Work around a strange bug wit reactive/juju 1.25.x
There is a strange bug where, randomly, and only with Juju 1.25.9 or 10,
the conversation for a relation doesn't exist, even though the relation
has been made.  This is a workaround that appears to have no impact on
using charm's functionality.

The clear_changed() methods are to remove a {relation-name}.changed flag
that gets set when the data on the relation changes.  However, not being
able to clear it, simply means that it will still be set the next time
the charm runs which might mean that some handlers run again.  This
shouldn't be a problem as, by design, reactive handlers MUST be
idempotent.

Change-Id: I26ec730d88e04fb3942d0240bf4eba6c6642a168
Related-Bug: 1663633